qlik_app/qlik/qlik-fastapi-backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
websocket-client==1.6.1
PyJWT==2.8.0
pandas==2.2.3
numpy>=1.26.0
matplotlib>=3.8.0
sqlalchemy
httpx
pydantic>=2.10.0
msal==1.31.0
reportlab>=4.0.0